Test scenario objects
The scenario objects works as a playlist of a test run. The scenario
object may be filtered according to object names and in the future
according to tags.

The scenario objects are responsible for creation of testing objects.
That's it, the execution framework takes one scenario item by one
and lets the item create the object and executes it.