Expression assertions
A new assertion function is offered. The OTest2 preprocessor parses
an expression passed as the first parameter and if it finds
a relational operator at the root of the expression it handles
the assertion as the same as the comparison assertions (comparator
is the template parameter and the operands are passed separately
into the assertion function).